refactor(core-flex-grid): break calculate reverse level into own file
